pyproject.toml
[project]
name = "cv-processing-2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.150.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
]
